+++ probe-rs/src/config/target.rs
@@ -1,10 +1,9 @@
 use probe_rs_target::{Architecture, ChipFamily};
 
 use super::{Core, MemoryRegion, RawFlashAlgorithm, RegistryError, TargetDescriptionSource};
-
-use crate::architecture::arm::sequences::nrf53::Nrf5340;
-use crate::architecture::arm::sequences::nxp::LPC55S69;
-use crate::architecture::arm::sequences::ArmDebugSequence;
+use crate::architecture::arm::sequences::{
+    nrf53::Nrf5340, nxp::LPC55S69, stm32::Stm32h7, ArmDebugSequence,
+};
 use crate::architecture::riscv::sequences::esp32c3::ESP32C3;
 use crate::architecture::riscv::sequences::{DefaultRiscvSequence, RiscvDebugSequence};
 use crate::flashing::FlashLoader;
@@ -99,6 +98,9 @@ impl Target {
         } else if chip.name.starts_with("nRF5340") {
             log::warn!("Using custom sequence for nRF5340");
             debug_sequence = DebugSequence::Arm(Nrf5340::create());
+        } else if chip.name.starts_with("STM32H7") {
+            log::warn!("Using custom sequence for STM32H7");
+            debug_sequence = DebugSequence::Arm(Stm32h7::create());
         }
 
         Ok(Target {